[iter2] Secure Tour Routes ...

+++ controllers/tourControllers.js
@@ -4,7 +4,8 @@ const mongoose = require("mongoose");
 // GET /tours
 const getAllTours = async (req, res) => {
   try {
-    const tours = await Tour.find({}).sort({ createdAt: -1 });
+    const user_id = req.user._id;
+    const tours = await Tour.find({ user_id }).sort({ createdAt: -1 });
     res.status(200).json(tours);
   } catch (error) {
     res.status(500).json({ message: "Failed to retrieve tours" });
@@ -14,7 +15,8 @@ const getAllTours = async (req, res) => {
 // POST /tours
 const createTour = async (req, res) => {
   try {
-    const newTour = await Tour.create({ ...req.body });
+    const user_id = req.user._id;
+    const newTour = await Tour.create({ ...req.body, user_id });
     res.status(201).json(newTour);
   } catch (error) {
     res.status(400).json({ message: "Failed to create tour", error: error.message });
@@ -30,7 +32,8 @@ const getTourById = async (req, res) => {
   }
 
   try {
-    const tour = await Tour.findById(tourId);
+    const user_id = req.user._id;
+    const tour = await Tour.findById({_id: tourId, user_id: user_id});
     if (tour) {
       res.status(200).json(tour);
     } else {
@@ -50,8 +53,9 @@ const updateTour = async (req, res) => {
   }
 
   try {
+    const user_id = req.user._id;
     const updatedTour = await Tour.findOneAndUpdate(
-      { _id: tourId },
+      { _id: tourId,  user_id: user_id },
       { ...req.body },
       { new: true }
     );
@@ -74,7 +78,8 @@ const deleteTour = async (req, res) => {
   }
 
   try {
-    const deletedTour = await Tour.findOneAndDelete({ _id: tourId });
+    const user_id = req.user._id;
+    const deletedTour = await Tour.findOneAndDelete({ _id: tourId, user_id: user_id });
     if (deletedTour) {
       res.status(204).send(); // 204 No Content
     } else {
